requirements.txt
requests
websocket-client
psutil
python-dotenv
